Autor, Livro and LivroInfo
Foram adicionados endpoints de autor, livro e livroinfo para implementar ainda mais a api.

+++ services/autor.service.js
@@ -1,29 +1,36 @@
 import AutorRepository from "../repositories/autor.repository.js";
+import LivroRepository from "../repositories/livro.repository.js";
 
 async function createAutor(autor) {
-    return await AutorRepository.insertAutor(autor);
-};
+  return await AutorRepository.insertAutor(autor);
+}
 
 async function updateAutor(autor) {
-    return await AutorRepository.updateAutor(autor);
-};
+  return await AutorRepository.updateAutor(autor);
+}
 
 async function deleteAutor(id) {
-    await AutorRepository.deleteAutor(id);
-};
+  const livros = await LivroRepository.getLivroByAutorId(id);
+  if (livros.length > 0) {
+    throw new Error(
+      "Não é possível deletar este autor pois há um livro cadastrado em seu ID."
+    );
+  }
+  await AutorRepository.deleteAutor(id);
+}
 
 async function getAutores() {
-    return await AutorRepository.getAutores();
-};
+  return await AutorRepository.getAutores();
+}
 
 async function getAutor(id) {
-    return await AutorRepository.getAutor(id);
-};
+  return await AutorRepository.getAutor(id);
+}
 
 export default {
-    createAutor,
-    updateAutor,
-    deleteAutor,
-    getAutores,
-    getAutor
+  createAutor,
+  updateAutor,
+  deleteAutor,
+  getAutores,
+  getAutor,
 };
